Mine Tune added
Added a new Tune to the game which spawns a mine behind the player and
detonates if the enemy gets within its trigger. The mine does not
detonate when the owner passes through its trigger. The mine also
disappears after 4 seconds if it is not detonated so the owner cannot
litter the ground with mines.

TODO: Color of the mines needs to be dependent on the color of the owner
so that it is apparent who the mine belongs to.